Some classes
pushing because rider keeps crashing

+++ C#/SEP3/Shared/domain/Purchase.cs
@@ -3,11 +3,11 @@ namespace Shared;
 public class Purchase
 {
     public List<OrderItem> orderItems { get; set; }
-    public int userId { get; set; }
+    public long userId { get; set; }
     public DateTime dateTime { get; set; }
 
 
-    public Purchase(List<OrderItem> orderItems, int userId)
+    public Purchase(List<OrderItem> orderItems, long userId)
     {
 
         this.orderItems = orderItems;
